backend/requirements.txt
# ── Core Flask ────────────────────────────────────────────────
flask==3.0.3
flask-cors==4.0.1
flask-jwt-extended==4.6.0
werkzeug==3.0.3
python-dotenv==1.0.1

# ── Anthropic ─────────────────────────────────────────────────
anthropic==0.28.0
httpx==0.27.0

# ── ML / RAG (Windows-safe versions) ─────────────────────────
faiss-cpu==1.8.0
sentence-transformers==3.0.1
numpy==1.26.4

# ── Document parsing ──────────────────────────────────────────
PyPDF2==3.0.1
pdfplumber==0.11.1
pandas==2.2.2
openpyxl==3.1.5

# ── Web scraping ──────────────────────────────────────────────
requests==2.32.3
beautifulsoup4==4.12.3

# ── Voice (optional - install separately if needed) ───────────
gTTS==2.5.1

# ── Export ────────────────────────────────────────────────────
reportlab==4.2.2

# ── Utilities ─────────────────────────────────────────────────
apscheduler==3.10.4
